+++ Sources/xcodeproj/XCConfigurationList.swift
@@ -6,7 +6,7 @@ public final class XCConfigurationList: PBXObject {
     // MARK: - Attributes
 
     /// Element build configurations.
-    public var buildConfigurations: [PBXObjectReference]
+    public var buildConfigurationsReferences: [PBXObjectReference]
 
     /// Element default configuration is visible.
     public var defaultConfigurationIsVisible: Bool
@@ -19,16 +19,27 @@ public final class XCConfigurationList: PBXObject {
     /// Initializes the element with its properties.
     ///
     /// - Parameters:
-    ///   - buildConfigurations: element build configurations.
+    ///   - buildConfigurationsReferences: element build configurations.
     ///   - defaultConfigurationName: element default configuration name.
     ///   - defaultConfigurationIsVisible: default configuration is visible.
-    public init(buildConfigurations: [PBXObjectReference],
+    public init(buildConfigurationsReferences: [PBXObjectReference],
                 defaultConfigurationName: String? = nil,
                 defaultConfigurationIsVisible: Bool = false) {
-        self.buildConfigurations = buildConfigurations
+        self.buildConfigurationsReferences = buildConfigurationsReferences
         self.defaultConfigurationName = defaultConfigurationName
         self.defaultConfigurationIsVisible = defaultConfigurationIsVisible
         super.init()
+    }
+
+    // MARK: - Public
+
+    /// Returns the build configurations.
+    ///
+    /// - Returns: build configurations.
+    /// - Throws: an error if the build configurations are not defined in the project
+    /// to which the configuration list belongs.
+    public func buildConfigurations() throws -> [XCBuildConfiguration] {
+        return try buildConfigurationsReferences.map({ try $0.object() })
     }
 
     // MARK: - Decodable
@@ -43,8 +54,8 @@ public final class XCConfigurationList: PBXObject {
         let objects = decoder.context.objects
         let objectReferenceRepository = decoder.context.objectReferenceRepository
         let container = try decoder.container(keyedBy: CodingKeys.self)
-        let buildConfigurationsReferences: [String] = try container.decode(.buildConfigurations)
-        buildConfigurations = buildConfigurationsReferences.map({ objectReferenceRepository.getOrCreate(reference: $0, objects: objects) })
+        let buildConfigurationsReferencesStrings: [String] = try container.decode(.buildConfigurations)
+        buildConfigurationsReferences = buildConfigurationsReferencesStrings.map({ objectReferenceRepository.getOrCreate(reference: $0, objects: objects) })
         defaultConfigurationIsVisible = try container.decodeIntBool(.defaultConfigurationIsVisible)
         defaultConfigurationName = try container.decodeIfPresent(.defaultConfigurationName)
         try super.init(from: decoder)
@@ -73,7 +84,7 @@ extension XCConfigurationList: PlistSerializable {
     func plistKeyAndValue(proj _: PBXProj, reference: String) throws -> (key: CommentedString, value: PlistValue) {
         var dictionary: [CommentedString: PlistValue] = [:]
         dictionary["isa"] = .string(CommentedString(XCConfigurationList.isa))
-        dictionary["buildConfigurations"] = try .array(buildConfigurations
+        dictionary["buildConfigurations"] = try .array(buildConfigurationsReferences
             .map { configReference in
                 let config: XCBuildConfiguration = try configReference.object()
                 return .string(CommentedString(configReference.value, comment: config.name))